+++ src/Kronika/Instant.php
@@ -88,10 +88,10 @@ final readonly class Instant
      *
      * ```
      * // 1767161730.004545
-     * $this->at(new \DateTimeZone('+01:00'));  // 2025-12-31 12:15:30.004545 +01:00
+     * $this->in(new \DateTimeZone('+01:00'));  // 2025-12-31 12:15:30.004545 +01:00
      * ```
      */
-    public function at(\DateTimeZone $timezone): ZonedDateTime
+    public function in(\DateTimeZone $timezone): ZonedDateTime
     {
         return ZonedDateTime::fromInstant($this, $timezone);
     }
@@ -370,6 +370,12 @@ final readonly class Instant
     public function __toString(): string
     {
         return (string)$this->number();
+    }
+
+    #[\Deprecated('use in() instead.', since: '0.4.2')]
+    public function at(\DateTimeZone $timezone): ZonedDateTime
+    {
+        return $this->in($timezone);
     }
 
     /** @internal */